finish HasMany function.
